// src/app/modules/auth/models/iuser.ts
export interface Iuser {
    E_Cityzen_ID: number;
    UserType: String;
    Name: String;
    CompanyName: String;
    EmailId: String;
    MobileNo: String;
    Password: string;
    HashMap: string;
    OTP: string;
    DOB: string;
    Gender: string;
    Road_Name: string;
    Locality_Name: string;
    District: string;
    State: string;
    Pin_Code: string;
    IDProof_Name: string;
    IDProof_Number: string;
    IDProofFile_Path: string;
    IS_Active: string;
}